ready manual scoring Healthy old control
completed manual scoring for all participants in the healthy old control group. Note that comments are included when a trial assignment was unclear - to be double check by another scorer. ALSO note that there were out-of-bound trials with NaN value for the real inferred angle - need to decide what to do with these.